Add files via upload
Beavers Get Cooler in Winter
Here is a chart of the body temperature data by hour for 2 beavers observed over 2 separate days in Wisconsin. This chart was run on December 22, 2016 in R version 3.3.2 using the plotly package and the beaver1 and beaver2 data in the datasets package.

The body temperature drops as the winter draws near.